generalize sulfur cube block model fix
