--- common/src/main/java/net/irisshaders/iris/mixin/entity_render_context/SulfurCubeMixin2.java
@@ -1,46 +1,0 @@
-package net.irisshaders.iris.mixin.entity_render_context;
-
-import com.llamalad7.mixinextras.sugar.Local;
-import com.llamalad7.mixinextras.sugar.Share;
-import com.llamalad7.mixinextras.sugar.ref.LocalIntRef;
-import com.mojang.blaze3d.vertex.PoseStack;
-import net.irisshaders.iris.mixinterface.SulfurCubeStateExtension;
-import net.irisshaders.iris.shaderpack.materialmap.WorldRenderingSettings;
-import net.irisshaders.iris.uniforms.CapturedRenderingState;
-import net.minecraft.client.renderer.SubmitNodeCollector;
-import net.minecraft.client.renderer.entity.SulfurCubeRenderer;
-import net.minecraft.client.renderer.entity.layers.SulfurCubeInnerLayer;
-import net.minecraft.client.renderer.entity.state.SulfurCubeRenderState;
-import net.minecraft.client.renderer.state.level.CameraRenderState;
-import net.minecraft.world.entity.monster.cubemob.SulfurCube;
-import net.minecraft.world.level.block.state.BlockState;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Unique;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-@Mixin(SulfurCubeInnerLayer.class)
-public class SulfurCubeMixin2 {
-	@Inject(method = "submit(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/SubmitNodeCollector;ILnet/minecraft/client/renderer/entity/state/SulfurCubeRenderState;FF)V", at = @At("HEAD"))
-	private void onRender(PoseStack poseStack, SubmitNodeCollector submitNodeCollector, int lightCoords, SulfurCubeRenderState state, float yRot, float xRot, CallbackInfo ci, @Share("lastBState") LocalIntRef ref) {
-		ref.set(CapturedRenderingState.INSTANCE.getCurrentRenderedBlockEntity());
-		iris$setupId(((SulfurCubeStateExtension) state).getBlock());
-	}
-
-	@Inject(method = "submit(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/SubmitNodeCollector;ILnet/minecraft/client/renderer/entity/state/SulfurCubeRenderState;FF)V", at = @At("TAIL"))
-	private void onRenderEnd(PoseStack poseStack, SubmitNodeCollector submitNodeCollector, int lightCoords, SulfurCubeRenderState state, float yRot, float xRot, CallbackInfo ci, @Share("lastBState") LocalIntRef ref) {
-		CapturedRenderingState.INSTANCE.setCurrentBlockEntity(ref.get());
-		CapturedRenderingState.INSTANCE.setCurrentRenderedItem(0);
-	}
-
-	@Unique
-	private void iris$setupId(BlockState state) {
-		if (WorldRenderingSettings.INSTANCE.getBlockStateIds() == null) return;
-
-		CapturedRenderingState.INSTANCE.setCurrentBlockEntity(1);
-
-		//System.out.println(WorldRenderingSettings.INSTANCE.getBlockStateIds().getInt(blockItem.getBlock().defaultBlockState()));
-		CapturedRenderingState.INSTANCE.setCurrentRenderedItem(WorldRenderingSettings.INSTANCE.getBlockStateIds().getOrDefault(state, 0));
-	}
-}
