+++ common/src/main/resources/mixins.iris.json
@@ -100,9 +100,8 @@
     "entity_render_context.MixinModelSubmit",
     "entity_render_context.MixinTextFeatureRenderer",
     "entity_render_context.MixinTextSubmit",
-    "entity_render_context.SulfurCubeStateMixin",
-    "entity_render_context.SulfurCubeMixin",
-    "entity_render_context.SulfurCubeMixin2",
+    "entity_render_context.BlockModelRenderStateMixin",
+    "entity_render_context.BlockModelResolverMixin",
     "entity_render_context.MixinCustomGeometrySubmit",
     "entity_render_context.MixinItemSubmit",
     "entity_render_context.MixinCustomFeatureRenderer",
